Logout log correct

+++ lib/provider/orders.dart
@@ -57,7 +57,7 @@ class Orders with ChangeNotifier{
               .map(
                 (item) => CartItem(
                       id: item['id'],
-                      price: item['price'],
+                      price: item['price'].toDouble(),
                       quantity: item['quantity'],
                       title: item['title'],
                     ),
